+++ examples/ibllib/ephys_qc_raw.py
@@ -9,14 +9,14 @@ import alf.io
 
 
 def _plot_spectra(outpath, typ, savefig=True):
-    spec = alf.io.load_object(outpath, '_spikeglx_ephysQcTime' + typ.upper())
+    spec = alf.io.load_object(outpath, '_iblqc_ephysSpectralDensity' + typ.upper())
 
     sns.set_style("whitegrid")
     plt.figure(figsize=[9, 4.5])
     ax = plt.axes()
-    ax.plot(spec['freq'], 20 * np.log10(spec['power'] + 1e-14),
+    ax.plot(spec['freqs'], 20 * np.log10(spec['power'] + 1e-14),
             linewidth=0.5, color=[0.5, 0.5, 0.5])
-    ax.plot(spec['freq'], 20 * np.log10(np.median(spec['power'] + 1e-14, axis=1)), label='median')
+    ax.plot(spec['freqs'], 20 * np.log10(np.median(spec['power'] + 1e-14, axis=1)), label='median')
     ax.set_xlabel(r'Frequency (Hz)')
     ax.set_ylabel(r'dB rel to $V^2.$Hz$^{-1}$')
     if typ == 'ap':
@@ -29,7 +29,7 @@ def _plot_spectra(outpath, typ, savefig=True):
 
 
 def _plot_rmsmap(outfil, typ, savefig=True):
-    rmsmap = alf.io.load_object(outpath, '_spikeglx_ephysQcTime' + typ.upper())
+    rmsmap = alf.io.load_object(outpath, '_iblqc_ephysTimeRms' + typ.upper())
     plt.figure(figsize=[12, 4.5])
     axim = plt.axes([0.2, 0.1, 0.7, 0.8])
     axrms = plt.axes([0.05, 0.1, 0.15, 0.8])
@@ -39,7 +39,8 @@ def _plot_rmsmap(outfil, typ, savefig=True):
     axrms.set_ylim(0, rmsmap['rms'].shape[1])
 
     im = axim.imshow(20 * np.log10(rmsmap['rms'].T + 1e-15), aspect='auto', origin='lower',
-                     extent=[rmsmap['times'][0], rmsmap['times'][-1], 0, rmsmap['rms'].shape[1]])
+                     extent=[rmsmap['timestamps'][0], rmsmap['timestamps'][-1],
+                             0, rmsmap['rms'].shape[1]])
     axim.set_xlabel(r'Time (s)')
     axim.set_ylabel(r'Channel Number')
     plt.colorbar(im, cax=axcb)
@@ -55,7 +56,8 @@ def _plot_rmsmap(outfil, typ, savefig=True):
 
 
 if __name__ == "__main__":
-    fbin = Path('/datadisk/Data/FlatIron/ephys/mainenlab/FC034/FC034_g0_t0.imec.lf.bin')
+    fbin = Path('/mnt/s1/Data/Subjects/ZM_1735/2019-08-01/001/raw_ephys_data/probe_left/'
+                '_iblrig_ephysData.raw_g0_t0.imec.ap.bin')
     ephysqc.extract_rmsmap(fbin)  # make sure you send a path for the time being and not a string
     typ = 'lf'
     outpath = fbin.parent